Remove extra macOS without-repo runs
It looks like the initial macOS without-repo run with 34 failures
was due to some mistake setting up the run. It's in the history of
this repo in case I need to look back at it, but I the results were
the same for all the other macOS without-repo runs.

This keeps the last of each of the two groups of macOS without-repo
runs (the use-archives and ignore-archives groups).